application/requirements.txt
beautifulsoup4
requests
pymongo
pyyaml
google-api-python-client
google-auth-httplib2
google-auth-oauthlib